Expand gfs_plots.yaml, add use of 'all' for channel list (#24)
This draft PR includes:

- addition of `instrument_channels.yaml` which contains the full channel
list for all radiance instruments. If `all` is specified in the
`[model]_plots.yaml` file then the channel list from
`instruement_channels.yaml` is used. Otherwise the list in
`[model]_plots.yaml` is used. This provides a means to limit the number
of channels plotted if the user is trying to dial in on a specific
issue.
- expanded (but not yet complete) `gfs_plots.yaml`. All radiance
time-series and summary plots are included. I'm still adding things to
this as I test the plot run-times. With all time series & summary plots
it has a wall time of ~1:45 now. I'm going to have to divide
`gfs_plot.yaml` into pieces to keep run times below 2 hrs.

I know this isn't near ready for release. But I'd like folks to take a
look at the implementation of `all` channels in `plotObsMon.py` and let
me know what you think. Thanks.