Add some of the control files to parm/cntrl (#29)
As per the discussion in #27 I've come around to a hybrid solution for
the control files. This PR sets up the mechanism for resolving a yaml
tag of `{{CONTRO}}` to the `obs-monitor/src/parm/cntrl` directory. This
PR puts 3 new (to obs-monitor) control files there -- 2 minimization
control files and the conventional histogram control file. These three
files do not contain any assimilation flags or other information that
will change over time. In the legacy MinMon and ConMon these files are
either stored in a parm directory or generated on-the-fly, so putting
them in the `obs-monitor/parm` directory is consistent with that use.



Closes #27