Fix minimization reduction plot file location, cycle time display (#31)
#30 (add missing reduction plot) was wrong (mostly). The reduction plot
was getting generated but it was not saved to the same location as the
minimization gnorm plots. That's now fixed.

Also I fixed an unrelated problem with the gnorm and reduction labels of
the cycle hour. Turns out `wxflow/jinja's` time filter will accept
strftime( [format string]) so conversion in the yaml file was simple.


Closes #30