Port obs-monitor to wcoss2 (#34)
This release allows obs-monitor to run on wcoss2. The strategy is to
split the model level yaml plot file into lots of pieces and submitted
as a single serial job with one node per piece (plot job). Satellite
plots (rad and ozn) are broken down by instrument, for iasi and cris
down to instrument/plot_type. The minimization and conventional plots
are submitted as two plot jobs without breaking them down further.

There are some caveats:

- Surface pressure plots are not working on either wcoss2 or hera. This
is a problem in eva (`mon_data_space.py`), and issue
[#194](https://github.com/JCSDA-internal/eva/issues/194) has been opened
to address it.
- `Logos `and `tight layout` options have been temporarily commented out
of the yaml plot tempates. These features fail on wcoss2 and seem to be
issues in emcpy. Issue
[#142](https://github.com/NOAA-EMC/emcpy/issues/142) has been opened to
address the problem.
- Some of the plots in `parm/gfs/gfs_plots.yaml` have been commented out
owing to what seems to be file corruption issues in my test data suite
(probably hera stmp space issues).

Testing has been conducted on wcoss2 and hera, and all plots work on
both machines within the stated caveats.