Port obs-mon to orion and hercules, add lua module files to create obs-mon python ve (#42)
This PR includes:

- Port Obs-Monitor to orion & hercules. 
- Addition of lua module files to create an obs-mon python virtual
environment. Note that I have the ve components temporarily installed in
my own noscrub space. I know this isn't suitable for a permanent
solution but thought it would be an acceptable placeholder.
- Correct use of CFP on wcoss2.  Really this time I've got it.  Honest.
- Removed file cleanup at finish. This needs to be done as a separate
job submission which runs after the plot job finishes. I'll open a new
issue to get that done.

Lastly, as describe in the discussion in #41 hercules is now having what
looks like communications issues when it runs eva. Execution hangs when
trying to resolve a url request from cartopy. That may be a system
issue, or it might be a permissions issue from the ve components which I
parked in my noscrub space. Though that location works fine on orion
(and wcoss2). I'm not quite sure how much to make of that.



Closes #41