Add cleanup job (#44)
This PR adds a cleanup script which is run as a separate job following
the successful completion of the plot job. It copies the plot file
directories to `$OM_PLOTS` (defined in `parm/OM_config`) and
conditionally (per `$KEEPDATA` flag) removes the working directory.

I fully expect we'll need to make many changes in the output storage
location and directory structure as well as add additional features to
the cleanup script. This release is an effort to set up the basic
cleanup infrastructure. This has been successfully tested on `wcoss2`,
`hera`, `orion`, and `hercules`.

Close #43