Set up input data (#46)
This PR adds a new `OM_data` class which is used by `plotObsMon.py `to
position the necessary data and control files required for a given plot.
This is more complex than it might seem because the legacy DA monitors
each handle their file naming uniquely. Additionally there are 4 data
directory formats supported. This includes the 3 that the legacy DA
monitor supports and a very simple structure which is only for use with
obs-monitor for development purposes. The directory structures are:

- obs-mon (development)
- glb-wkfl (global workflow output)
- gfs-ops (gfs operational output)
- leg-mon (legacy DA monitors internal storage)

Note that the `plotObsMon.py` is written in a fairly procedural manner.
That's on me, and perhaps made some sense when it was pretty simple but
I think it was more a matter of my habit than the right way to go about
things. This PR begins to change that -- the OM_data class is written in
an OO fashion and I do plan to update plotObsMon.py to make it more OO
in approach in another PR.

All the template files in `parm/templates` have been modified to make
the path to the required files simply `{{DATA}}`, removing the following
sub-directory (i.e. `{{DATA}}/rad_data/`). All the path extensions are
now handled by OM_data as part of the directory structure support
mentioned above.


Closes #45