Add loop control to simplify yaml templates (#48)
This is the first cut at simplifying yaml templates with loop controls.
I've condensed the 'PDATEm*' values into a list and then iterate over it
to generate the necessary file names.

I had hopes of generating the PDATEm* list dynamically, but after
bashing my head against it (and ChatGPT's hallucinations) for the past 3
days I'm not convinced that is possible. The next step to pursue a
solution like that, I think, is to create a clean jinja2 environment
apart from wxflow's extension and see if developing such a solution is
possible. Within the wxflow's additions I can't generate any useful
debug information -- the debug attempts result in parsing exceptions and
no useful output.

So this draft is mostly a status check. If the changes to the one file
here is an acceptable solution I can change the other yaml template plot
files accordingly. If I should continue pursuing dynamic generation of
the list, then I will do that.

Closes #47 (some day).