Avoid attempting to make plots which have an incomplete specification (#50)
Avoid attempting to make any plots which have an incomplete
specification in the top-level yaml plot file. This is in response to
problems Andrew found in testing. Absent these changes such incomplete
specifications cause a catastrophic failure of all remaining plots. With
these changes incomplete specifications are skipped over and an
identifying warning message is produced.

Closes #49