Update obs-monitor wcoss2 module. (#54)
I ran into some problems updating the obs-mon ve on wcoss2 due mostly to
existing issues. These are fixed and I've updated the python version as
well. Small changes were required in the `wcoss2.lua` module to complete
the ve update.

The issue that prompted this fix was that plotting logos still didn't
work on wcoss2 despite the recent update to eva. With this change and
the rebuilt ve logos (and tight layout) now work on wcoss2.

Closes #53