Add Obs-monitor website project code (#64)
Initial commit of website components including React, Vite, and tailwind
related files. These are all the files necessary to load into VS Code
and generate the required website components. The eventual (actual)
website components are placed in a obsmon/dist/ directory which is
intentionally **not** included in this PR.

This is, regrettably, a large PR of > 100 files. Most of it won't make
sense. I _generally_ know what I'm working with now but only because
I've been hacking away at this for the past month. For review purposes I
suggest focusing on the README file and the src/*.jsx and
src/components/*.jsx files. The .jsx files are the React coded files
which are fed into node.js to generate the user interface. They contain
html, javascript, and react coding elements.

I also note I've hit some sort of CI failure upon pushing this code to
my fork so I'll have a look at that. Yippee.

Closes #63

---------

Co-authored-by: Ed Safford <1392868069113305@mil>