Make menu cascade .json file driven, add built website to repo (#70)
This PR makes a number of changes to support a user-modified (or
scripted setup) website possible without building via jsnode|React|vite
and an IDE. The number of changes is large but straightforward. They
include:

- Rename the `obs-monitor/web` directory to `obs-monitor/UIdev` to make
clear it's **not** the website directory.
- Move all the menu cascade component `.js` files from `src/data` to
`public/data` and convert & rename to `.json` files. The contents of the
`public/data` directory are transferred at build to the eventual
`website/data` directory without modification. They are loaded when the
site is accessed and the menu cascade constructed on-the-fly. This
enables a user's website to reflect the contents of their experiment
without rebuilding the site.
- Rename `UIdev/dist` directory to `UIdev/website` to indicate this
**is** the website files and code. While voluminous it doesn't contain
any problematic files (.tar, binaries, etc) for git. The contents of
this directory are what will be deployed for user websites.

Additional changes will likely be necessary as we settle on naming
conventions and data and image directory structures.

A follow-on issues will be generated shortly to automate and customize
website generation and provide user instructions for website creation.

Closes #68

---------

Co-authored-by: Ed Safford <1392868069113305@mil>