Add configuration file to customize the splash page (#73)
Add `configIndex.json` file and modify `Home.jsx` to enable
customization of the splash page at load time without rebuilding the
`.jsx` code.

Note that there are intentionally two copies of `configIndex.json`. One
is `UIdev/ObsMon/public/data/configIndex.json`. This is the stock
version used mostly as a placeholder for the UI building code. The
contents of `public/data` are copied without change to
`UIdev/ObsMon/website/data` during the site build. It's the `website`
directory contents that are deployed to the server and it is this
version that will get replaced with a customized version of
`configIndex.json` per the users' input/editing.

Incidental findings. Our server really doesn't like files named
`config.*`. Initially `configIndex.json` was named `config.json` and
this would _not_ load. It would be found by the fetch operation so not a
404 condition, but rather just hang with no explanation. Eventually
ChatGPT volunteered that 'yeah, some servers don't like config as a
filename, particularly government systems'. Ah, thanks, only spent most
of the morning on that one point so all good I guess. :(

Closes #72

---------

Co-authored-by: Ed Safford <1392868069113305@mil>