Add eva driver script (#62)
This PR introduces the first implementation of the infrastructure
required to run a driver script for observation monitoring. The current
version includes the following functionality:

- Copies NetCDF files from a specified experiment_dir to a runtime
directory.

- Dynamically generates an EVA configuration file using a Jinja2
template.

- Executes the EVA process with the generated configuration.

The copy-to-output-directory logic is currently commented out until we
make a decision on where we want to point these figures. We will
eventually support multiple templates to for all the figures we want to
generate.

Some paths and parameters are still hardcoded and would benefit from a
more flexible configuration or templating approach. Feedback and
suggestions on how to improve this are welcome.